rar5: fix 32-bit truncation in EX_UOWNER bounds check
On 32-bit platforms, (size_t)*extra_data_size truncates the int64_t
value when it exceeds SIZE_MAX (4 GiB).  Use (uint64_t) casts on both
sides of the comparison so the check is exact on all platforms.

Addresses review feedback on PR #3332.

+++ libarchive/archive_read_support_format_rar5.c
@@ -1547,7 +1547,7 @@ static int parse_file_extra_owner(struct archive_read* a,
 		 * this field. Rejecting an oversized length here also avoids
 		 * requesting a huge allocation from read_ahead() below. */
 		if(*extra_data_size < 0 ||
-		    name_size > (size_t)*extra_data_size) {
+		    (uint64_t)name_size > (uint64_t)*extra_data_size) {
 			archive_set_error(&a->archive,
 			    ARCHIVE_ERRNO_FILE_FORMAT, "Owner name is too long");
 			return ARCHIVE_FATAL;
@@ -1575,7 +1575,7 @@ static int parse_file_extra_owner(struct archive_read* a,
 			return ARCHIVE_EOF;
 
 		if(*extra_data_size < 0 ||
-		    name_size > (size_t)*extra_data_size) {
+		    (uint64_t)name_size > (uint64_t)*extra_data_size) {
 			archive_set_error(&a->archive,
 			    ARCHIVE_ERRNO_FILE_FORMAT, "Group name is too long");
 			return ARCHIVE_FATAL;
